span,diag: new span and diag component
