+++ cli/BUILD.bazel
@@ -8,6 +8,7 @@ rust_binary(
     deps = [
         "//bir",
         "//birgen",
+        "//diag",
         "//ast",
         "//lexer",
         "//session",
