+++ diag/BUILD.bazel
@@ -1,11 +1,10 @@
 load("@rules_rust//rust:defs.bzl", "rust_library")
 
 rust_library(
-	name = "session",
+	name = "diag",
 	srcs = glob(["src/*.rs"]),
 	deps = [
-		"//diag",
-		"@crates//:anyhow",
+		"//span",
 		"@crates//:annotate-snippets",
 	],
 	visibility = ["//visibility:public"],
