+++ span/src/lib.rs
@@ -1,0 +1,11 @@
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
+pub struct SourceSpan {
+    pub start: usize,
+    pub end: usize,
+}
+
+impl SourceSpan {
+    pub fn new(start: usize, end: usize) -> Self {
+        Self { start, end }
+    }
+}
